piko: integrate agentv2 and serverv2
Integrates agentv2 and serverv2 into main.

This contains a complete rewrite of the agent to server protocol that
sends traffic as TCP over WebSocket rather than sending individually
encoded HTTP requests.

The simplifies how both the agent and server proxy requests as the
'connection' between agent and server is a net.Conn.